Begin header file refactoring.  This is a work-in-progress.
