+++ crypto/include/auth.h
@@ -46,9 +46,9 @@
 #ifndef AUTH_H
 #define AUTH_H
 
+#include "srtp.h"
 #include "datatypes.h"          
 #include "err.h"                /* error codes    */
-#include "crypto.h"		/* for auth_type_id_t */
 #include "crypto_types.h"	/* for values of auth_type_id_t */
 
 typedef struct auth_type_t *auth_type_pointer;
