+++ crypto/include/cipher.h
@@ -46,10 +46,10 @@
 #ifndef CIPHER_H
 #define CIPHER_H
 
+#include "srtp.h"
 #include "datatypes.h"          
 #include "rdbx.h"               /* for xtd_seq_num_t */
 #include "err.h"                /* for error codes  */
-#include "crypto.h"		/* for cipher_type_id_t */
 #include "crypto_types.h"	/* for values of cipher_type_id_t */
