+++ srtp/srtp.c
@@ -42,8 +42,10 @@
  *
  */
 
-
+#include "srtp.h"
 #include "srtp_priv.h"
+#include "crypto_types.h"
+#include "err.h"
 #include "ekt.h"             /* for SRTP Encrypted Key Transport */
 #include "alloc.h"           /* for crypto_alloc()          */
 #ifdef OPENSSL
